% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{a_class}
\alias{a_class}
\title{Assert input object class}
\usage{
a_class(object = NULL, class = NULL, allow_null = FALSE)
}
\arguments{
\item{object}{Any object to be tested for.}

\item{class}{A character scalar.}

\item{allow_null}{A logical scalar. Is NULL allowed ?}
}
\value{
same as input
}
\description{
Assert input object class
}
