% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{a_count}
\alias{a_count}
\title{Assert input is a positive integer scalar (count)}
\usage{
a_count(count = NULL, allow_null = FALSE)
}
\arguments{
\item{count}{A positive integer scalar.}

\item{allow_null}{A logical scalar. Is NULL allowed ?}
}
\value{
same as input
}
\description{
Assert input is a positive integer scalar (count)
}
