% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{a_dir}
\alias{a_dir}
\title{Assert input is a writable directory}
\usage{
a_dir(dir = NULL, allow_null = FALSE)
}
\arguments{
\item{dir}{A charactor scalar of directory.}

\item{allow_null}{A logical scalar. Is NULL allowed ?}
}
\value{
same as input
}
\description{
Assert input is a writable directory
}
