% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{a_flag}
\alias{a_flag}
\title{Assert input is a logical scalar (flag)}
\usage{
a_flag(flag = NULL)
}
\arguments{
\item{flag}{A logical scalar.}
}
\value{
same as input
}
\description{
Assert input is a logical scalar (flag)
}
