% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{a_str}
\alias{a_str}
\title{Assert input is a character scalar (string)}
\usage{
a_str(string = NULL, allow_null = FALSE)
}
\arguments{
\item{string}{A character scalar.}

\item{allow_null}{A logical scalar. Is NULL allowed ?}
}
\value{
same as input
}
\description{
Assert input is a character scalar (string)
}
